+++ src/components/Header/Header.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
 import { Container, Logo, NavLink } from './Header.style';
+import { Link } from 'react-router-dom';
 
 const Header = () => {
   return (
