+++ src/helpers/getHourlyIntervalsPerDay.js
@@ -3,16 +3,16 @@ import isSameDay from 'date-fns/isSameDay';
 import subSeconds from 'date-fns/subSeconds';
 
 import { getEpoch } from './getDateInterval';
-import generate24HrPostTimes from './getGridTimesIntervals';
+import generate24HrPostTimes from './getGridHeaderInterval';
 import lastWeekDays from './getDaysOfTheWeek';
 
-//where hour type = 1 is is the first hour of an interval, and 2 is the second, hour type?
 const getHourlyIntervalsPerDay = (increment = 2, day) => {
   const hourlyIntervals = generate24HrPostTimes(increment, true);
 
+  var hourToHourIntervals = [];
+
   let start = 0;
   let end = 1;
-  let hourToHourIntervals = [];
   let startTime = lastWeekDays[day][0];
   let endTime = lastWeekDays[day][1];
 
@@ -45,6 +45,7 @@ const getHourlyIntervalsPerDay = (increment = 2, day) => {
   }
 
   const lastInterval = hourToHourIntervals.length - 1;
+
   if (
     hourToHourIntervals[lastInterval] &&
     hourToHourIntervals[lastInterval]['10:00pm']
